requirements.txt
streamlit
requests
pandas
fpdf2
streamlit-autorefresh
deep-translator